Expose an API for finer control over heap allocator blocking behavior.
The current `timeout` argument of `heap_allocate*` functions determines
whether or not and how long the allocator should block for an allocation
to be fulfiled. However it does not control on what events the allocator
should block.

Currently the allocator would block for three reasons: 1) there is no
heap space available anymore, 2) the quota is exhausted, and 3) the
quarantine is holding back memory which could be used for the
allocation.

For the network stack use-case, we need a finer way to control this
blocking behavior, for example to only wait on the quarantine.

This commit introduces such an API through a new `flags` argument.

Add tests for this new API, but not for `AllocateWaitQuotaExceeded` as
it requires manipulating the quarantine to be done reliably.

Update the stack usage of some allocator functions which has
correspondingly increased.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>