Use @jrtc27's approach to enable large sealed objects.
Allocating large sealed objects is difficult because sealed objects have
a header that contains the virtual sealing type and both the full
(sealed) object including the header and the (unsealed) part with the
header excluded must be representable.

@jrtc27 pointed out that there's no real need for the header to be at
the start of the allocation.  This neat trick lets us lay out the
objects as:

```
|Padding|Header|Unsealed Object|
        ^
        Address of sealed capability is here.
```

As long as we first round up the size of the unsealed object to be
representable, we can always then create a larger allocation that
contains padding at the start to preserve representability of the sealed
object.

This removes the restriction that we can't have sealed objects larger
than 4072 bytes, which will eliminate some annoying redirection in a few
places.

Fixes #242