Add APIs to retrieve the thread ID of the owner of a lock.
It is commonly needed in error handlers to have to check if the thread
on which the error handler was invoked was holding a particular lock
when it crashed. Typically, we would want to do this to be able to
forcefully release the lock. This is currently difficult to do, as the
thread ID which is stored in priority inheriting flag locks is not
exposed through the C/C++ APIs.

This commit adds news APIs `get_owner_thread_id` (C++) and
`flaglock_priority_inheriting_get_owner_thread_id` (C) to retrieve the
thread ID of the owner of a lock.

Following this commit, error handlers can call `get_owner_thread_id` (or
`flaglock_priority_inheriting_get_owner_thread_id`) on a lock and
compare the result with `thread_id_get` to know if the thread which
crashed owned the lock.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>