Test and document the linked list from `ds::linked_list`.
Although our linked list implementation does feature comments, they are
more about the underlying technical details and do not really help
newcomers understand how to use it. This matters because we do not have
`std::list`.

Address this by adding a test for the linked list which also explains
how to use it for a basic use-case.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@scisemi.com>