Add initial support for pre-shared globals. (#283)
This provides a mechanism to statically define globals that can be
shared between compartments statically. As with MMIO imports, these can
be imported with restricted permissions.

They currently show up in the audit log as MMIO imports. This is not
ideal, we should plumb their names through into the audit log so that
cheriot-audit can check that all references to a pre-shared object refer
to exactly one object.

Fixes #282