Fix a bug in the Sonata Ethernet driver.
The returned Ethernet frame was precisely bounded, but Ethernet frames
can be all sorts of odd sizes and may not be representable with the
capability bounds encoding.  This resulted in the tag being cleared,
which caused a trap in the firewall.  The firewall is designed to fail
closed, so this then prevented any further packets from arriving.

The bounding is purely defence in depth, so it's fine to use imprecise
bounds here.