Sonata Ethernet driver cleanups.
The log packet filtering was being wrapped in a conditional and then
using a template that used a different conditional, so never reported
anything.  Fix this to instantiate the template properly.

Also add two defines that tell FreeRTOS+TCP the offload features that
this network interface exposes.  This improves performance and reduces
code size.