Add Sonata ADC driver
This commit adds a simple Analogue-to-Digital Converter (ADC) driver for
Sonata. For now this only exposes a small subset of the XADC
functionality so that software is able to read XADC registers, modify
the clock speed, and power down the ADC to reduce power draw. The driver
supports extension to add and expose more features as is needed.