Fix atomicity of Sonata ethernet driver.
The network stack expects to be able to receive frames in one thread and
send them from another.  The Sonata Ethernet has some MMIO state that
is shared between the two.

This is not a complete fix because other SPI devices can also be
interleaved.  We see this sometimes when packets end up written into the
LCD display.