Add support for stackless error handlers.
These are given a stack bounded it was on entry into the compartment,
with an undefined offset.  As such, they can be used even when the cause
of the error is stack overflow.

Unlike the existing error handler, this cannot take a snapshot of the
register file, because there is no space for it.

Both kinds of error handler can be present.  When both are present, the
existing kind will be used if there is sufficient stack space to invoke
it, otherwise the stackless one will be invoked.

This also adds a 'library' that provides an implementation of a
stackless error handler that interoperates with the new setjmp-based
unwind code.  When an error occurs, it pops the top jump buffer and
invokes it.

A few other small cleanups while I was touching adjacent things:

 - There were some issues in the switcher API to recover the initial
   stack, the tests were passing largely by coincidence.  This code is
   now fixed and the tests improved.
 - A load of recently added libraries were not linked to in the README
   for libraries.  They are now.
 - Reserve the 16 bytes at the top of the stack on thread entry as well
   as on cross-compartment call.

Fixes #298