Disable including the ELF headers in the firmware image.
By default lld will try to write a copy of the ELF PHDRs in the final
memory image before the first loadable segment. Normally this is elided
because there is insufficient space before the first loadable segment,
but if the starting address specified for the instruction_memory (in
the board JSON description) leaves sufficient space this can happen. To
disable this behaviour add the -n option to the lld command line;
the other effects of -n are irrelvant: no page alignment of sections
(done already by the linker script), and link against static libraries.

Change-Id: I2b8e21c1d43dd66318f16d54b4c2c1babac337d3