Capability.bounds().set_inexact_at_most() fixes (#312)
- Don't take bounds by reference (rely on the caller to get the
resulting length afterwards) for consistency with the other APIs here.

- We were not falling into the bit-twiddling path in the case that the
length could not fit in a mantissa, and so could return an untagged
result. Instead of trying to do all the requisite logic to see if
setting bounds would succeed, just try to set the bounds and see if it
worked.

- Update tests to reflect new interface, and add another that would have
caught the above bug.

- Move the bit math out to a separate function, giving the compiler some
more flexibility in inlining.