Separate static and dynamic sealing types.
This separates the sealing hardware sealing types used for
heap-allocated objects from the sealing type used for static sealed
objects.  The main goal of this was to remove the allocator from the TCB
for statically sealed objects but, as a side benefit, we also delete
about 200 lines of code and make the binaries around 100 bytes smaller.

We had previously reserved a type for the scheduler's sealed types to
remove the allocator from the scheduler's TCB for static objects.  This
is no longer required.  The allocator is in the scheduler's TCB for
dynamically sealed objects (but it is anyway, because they live on the
heap), but not for statically sealed objects.  This removes a load of
special-case code from the scheduler.

This also means that we didn't increase the number of hardware sealing
types used, because the scheduler no longer needs its own one.

The existing APIs are preserved, but the token unsealing APIs now have
variants that will work *only* on static or dynamically sealed things
and so users can remove allocator from the TCB for other objects that
should be only statically sealed (and auditable), not heap allocated.