Remove space reserved at the start of privileged compartments.
We originally reserved one capability of space at the start of each of
the core RTOS compartments for a sealing key.  Then, when we added the
virtualised sealing mechanism to the allocator, this grew to two.

The allocator is the only compartment that uses both slots and, after
merging #316, the scheduler doesn't use any.  This space is therefore
wasted.

This commit removes this space and adds a mechanism that allows
privileged compartments to reserve one or two slots.

Eventually we can probably replace this with a normal imported sealing
capability, but for now this at least removes the wasted space.

This does not currently reduce code size, because replacing the inline
assembly with a load of a global adds a csetbounds.  This can be fixed
in the compiler independently.

While we're here, improve a test that could spuriously pass if the inner
compartment faulted.