switcher: Rework commentary
Include "Register Atlas"es and updates at key points, making it
hopefully easier to understand what's going on at any given point in the
code.

Annotate arcs in the CFG at their target, not just their source, as well
as IRQ disposition and live-in and live-out register sets.

The intention is that much of this should be amenable to automatic
verification, but no such tooling yet exists.

To ease review, this large commit changes no bytes in the actual
assembled output.

Co-authored-by: David Chisnall <github@theravensnest.org>
Co-authored-by: Murali Vijayaraghavan <vmurali@csail.mit.edu>
Co-authored-by: Robert Norton <robert.norton@microsoft.com>