Simplify message queues.
My original implementation was a bit CHERI-happy and put a lot of
capabilities in each endpoint to enforce properties with a
not-very-well-thought-through thread model.

The new code is simpler in a few ways:

 - The queue metadata and ring buffer are a single allocation.
 - The producer / consumer counters are indexed directly, not via
   indirection.
 - We don't do permission checks and claims for defensive programming in
   the send and receive functions, we use the unwind error handling so
   don't need to enumerate possible badness.
 - The library returns a pointer to the allocation directly, rather than
   a (large) value type and a pointer that is used for freeing.
 - Restricted endpoints are just a simple indirection layer and are only
   in the sealed version where they might make sense with a threat
   model.
 - The FreeRTOS compatibility APIs no longer need a heap allocation to
   store a pointer to another heap allocation.

This is an API break, but the new APIs are simpler so it's probably
worth it, and good to do before 1.0.  I believe the only users are using
the FreeRTOS-Compat wrappers, which are not changed (though can now
easily be extended to support message queues between compartments in the
compat layer).

Fixes #309