switcher: inline and prune check_compartment_stack_integrity
Inline check_compartment_stack_integrity since it's used exactly once
and makes some assumptions (and had a bug, hard-coding `csp` when it
meant `\reg`).

Remove the `csp`-relative `lb`, as we have already checked the things
that this would that the surviving subsequent permissions check would
not.  Leave the permission and alignedness checks.