switcher: ensure that MTCC is really NULL on fault
As @davidchisnall points out, just because sp's address is zero doesn't
mean that the rest of the cap is too.  So, if it looks like we're dead
in the water, make sure we're actually dead by zeroing mtcc completely.

While here, slightly tidy the code by flipping the condition and moving
the wedging operation out of line.  Add an instruction guaranteed to
trap after nulling MTCC.