sonata: simplified the PWM driver
There are now 6 general purpose PWM outputs and one dedicated to the
LCD's backlight. Instead of implementing the indexing logic and checks
in the driver. The driver is changed to only over map to a single PWM
output. We simply alias an array of these for the general purpose PWM
memory region.

The dutyCycle <= period assertion was removed, because dutyCycle >
period is the only way to acheive a constant high, 100% duty cycle
output from the Sonata system's PWM block.