lints: Removed v0.2 versions of the sonata drivers
There are name collisions between the old v0.2 versions of the of sonata
system drivers and the new v1.0 versions. These cause clang-tidy lints
to fail.

The v0.2 drivers are frozen; they should not change before being
depreciated and removed from the repository. It is therefore safe to
remove them from the lints.