Sonata simulator run script added
It uses the SONATA_SIMULATOR, SONATA_SIMULATOR_BOOT_STUB and
SONATA_SIMULATOR_UART_LOG environment variables. It has two modes, one
where the simulator runs in the foreground indefinitely and needs the user
to manually stop the script. The second mode runs the simulator in the
background and the UART log is monitored for a value given as a
parameter to the script. When running in the monitoring mode, there is a
60 second timeout just so the script does not run forever.

Example command, which returns 0 on success and 1 on failure:
scripts/run-sonata-sim.sh examples/01.hello_world/build/cheriot/cheriot/release/hello_world scripts/model_output/sonata-simulator/examples/01.hello_world.txt