Add lock guard steps and conditions.
Add two new constructors to the `LockGuard` class.

- Make it possible to specify a `step` argument: instead of attempting
  to acquire the lock with the timeout in one chunk, try to acquire the
  lock in a loop of steps of duration `step`, until the timeout expires.

- Make it possible to specify a `condition`: stop attempting to acquire
  the lock if the condition becomes `true`, checked at every `step`.

Both are necessary to improve the reliability of compartment reset for
the TCP/IP stack.

Specifically, steps allow us to handle the situation where a crash makes
us lose the reference to a lock on which a thread is blocked.  With
steps we are guaranteed that the thread will become live again in a
bounded amount of time (`step`), and try to re-acquire the lock: we can
thus free the lock through heap-free-all, and have the "lost" thread
crash when re-trying to acquire the lock.

However, it is not always possible to heap-free-all to free the lock,
e.g., if the lock was allocated with a caller's quota. In this case, we
can call the lock with a condition that checks if a reset is ongoing
(very simply through checking a bit in the state machine word) to
prevent the "lost" thread from blocking again.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@ubc.ca>