CI runs all examples on Sonata simulator
It add Sonata pre-release to the matrix of boards. Xmake run tries to
use the Sonata FPGA which does not exist in this CI.
The test suite doesn't run on the simulator at the moment because
HyperRAM is not supported in the RTOS right now. The
Sonata smoke test runs each example and checks the content of the UART
log against the model output.

Debug executables are currently not running on Sonata because the UART
log gets too full and causes the UART log to become unavailable inside
the GitHub runners.