Export freestanding symbols as unmangled.
Mangling these is painful for the compiler and we've missed at least one
place where it needs to happen.  We don't mangle the atomics, so it's a
good idea to also not mangle the freestanding ones.

As a transitional measure, export the functions as both their mangled
and unmangled forms.

The next step is to make the mangled form the compatibility definition
and use `asm` in the header to export these as non-mangled by default
for direct calls, then finally to remove `compat.S`.

Once this is merged, the compiler is free to start emitting `mem*` as
libcalls without name mangling.

Fixes #368