isa-test: extend the store test to check memory after store
It's good to check that memory wasn't modified in faulting cases and we stored the expected data in non-faulting cases.
One of the tests is redundant after an ISA change: it makes no difference whether the tag of a local capability being stored is set because it should will be cleared anyway.
Previously it would fault only if the tag was set.