Partial fix for #381: exit the simulator with failure even if the error handler causes a stack overflow.
This uses a stackless error handler to recover from stack overflow in
the stack-hungry error printing path. The handler just continues to
the simulator_exit call. This is not a very satisfactory solution as
it requires a call to switcher_handler_invocation_count_reset which
could result in an infinite loop if an error occurs in the handler
that is not handled by compartment_error_handler instead of by the
stackless error handler.

The stackless error handler will only be invoked if the compartment
links with the unwind_error_handler library.