Hardware revoker: correct wait for indicated epoch
The commentary says that we wait for the indicated (even) epoch to have
finished, but in practice we were merely waiting for the next one to
begin.  Because has_revocation_finished_for_epoch() had the intended
semantics and is the actual function used to gate release from
quarantine, wait_for_completion()'s incorrect waiting was merely
ineffecient, making malloc_internal spin a bit in the "Quarantine has
enough memory to satisfy allocation" case.