Improve futex_wake behaviour.
The current code yields and schedules a new thread if any equal or
higher-priority thread has become runnable.

The new code decomposes these cases.  If a new thread is runnable and
should run now, the current thread yields.  If a new thread is runnable
but is the same priority as the current thread, the scheduler ensures
that a timer interrupt is ready but does not yield immediately.

This improves the test suite performance by around 2% and a
producer-consumer microbenchmark by about 35%.