Adopt compiler support for sealed capabilities
This still supports the old mode and, because some existing code may
require use of SObj, provides a compatibility mode for migration.

With the sealed branch of the compiler, there is a __sealed_capability
qualifier.  This qualifier is part of the type system and sealed
capabilities cannot be dereferenced (the hardware enforces this, now the
compiler does as well).  The seal and unseal builtins are perform the
conversion between a `T*` and a `T*__sealed_capability`.