Add support for levels in the ConditionalDebug class.
The class is now instantiated with a threshold level, from None to Error
(for backwards compatibility, true means Information and false means
None).

Log messages also provide a log level (Information if none is provided)
and will fire if the log level is greater than or equal to the
threshold.

The build system now has a `debugLevelOption` that allows specifying
fine-grained levels.  The existing `debugOption` works as before.

The allocator is now extended to use `debugLevelOption`.