allocator-test: preflight slightly more explicit
The test of `p`'s validity after free() might be of a value spilled and
reloaded from the stack (either by us or by the cross-call path).  While
that's absolutely worth testing, also explicitly probe at the return
registers and `p` from assembler.  The compiler's understanding of the
calling convention should take care of this -- explicitly clobbering a1
with zero -- but make sure that it isn't a still-live copy of the
pointer we passed in.

While here, manually flush quarantine in hopes of ensuring that we don't
accidentally trigger the revoker with the free of `p`, so that we really
are probing at the load barrier's behavior.