Reduce scheduler permissions slightly.
The scheduler doesn't require store permission for futexes when it does
futex_wake, so don't require any permissions here.  This means that you
can spuriously wake any thread waiting on a futex that you have any
pointer to, even a read-only one, but the API permits spurious wakes.

The benefit from this is that the scheduler cannot tamper with a futex
word.  For example, if a futex is a trivial lock where 1 indicates locked
and 0 indicates unlocked, the scheduler currently can both wake waiters
and set the lock value to 0 making two threads believe that they
acquired the same lock.  With this change, an attacker who gets
arbitrary code execution in the scheduler cannot do this, they can only
cause threads to wake, see that the lock is already acquired, and then
go back to sleep.

The C++ atomics (used for the locks library) now enforce this least
privilege when calling the scheduler.

These changes increase the size of the test suite by 32 bytes.