New external IRQ latency benchmark
This uses existing peripherals that can be used as IRQ "loopbacks",
which fire some appreciable delay after a software event, to measure
cycle, instruction, or HPMC counters of interest.  Presently, the Ibex
revoker engine or the Sunburst UART1 can be used as triggers; it should
be easy to add other sources.