sdk/xmake: fix scheduler stack usage checks
The scheduler compartment must be built with the
"cheriot.component-stack-checks" rule, as with the allocator.

It would be nicer if we could do this from the stackCheckOption
invocation, but I don't think that's possible.  (Doubly so because the
scheduler's target name is computed.)

Bugfix on 579fc322ac31379271c1ac357b50eff1cd5e5eb9