debug.hh: use 'enum struct DebugLevel'
Prevent automagic conversion of DebugLevel to a boolean, because
DebugLevel::None is nonzero (in fact, the largest of the DebugLevel
values) as it is meant to be used as a threshold for reporting.

The use of an 'enum struct' means that we can't mistakenly write things
like "if constexpr (DEBUG...)" and end up with debug code compiled in
when DEBUG... is DebugLevel::None (or wonder why debugging might be
elided when it's ::Information).

Update the sole user of DebugLevel, the allocator, to compile.