Add a stream implementation and most of the FreeRTOS compat API.
The core of this is a pair of APIs for sending and receiving more than
one element to a queue in a single call.  These are also exposed for the
queue compartment, which lets you amortise the cost of cross-compartment
calls.

A stream is a message queue where queue items are one byte long.
Sending multiple queue items at a time gives a byte-stream abstraction.

FreeRTOS streams are a bit more complicated than CHERIoT RTOS version
and have explicit thresholds and callbacks.  These are approximated
here.