Sunburst: Fix Ethernet ICMP checksum field offset
See Table 4-35 (p. 47) of the datasheet for the KSZ8851SNL/SNLI
KSZ8851 SPI Ethernet MAC used on Sonata.

The ICMP Checksum Generation field should be the 8th (zero-indexed) bit
of the Transmit Control register, not the 9th. This meant that in the
current Ethernet driver, ICMP checksums are not generated by Sonata.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>