Softfloat implementation and smoke test.
This is divided into multiple libraries so that it's easy to pick just
the operations that you need.  There are four convenience targets:

 - softfloat32 (almost everything for `float`)
 - softfloat64 (almost everything for `double`)
 - softfloat (almost everything)
 - softfloatall (everything, including complex numbers)

Most of the time, one of the first three will be adequate.

The total code comes in at a bit under 15 KiB.
Just 32-bit float support (no complex numbers, no powers) comes in at
under 4 KiB.