Add support for printing floats and doubles.
This uses the float-to-string algorithm from AduinoJson, which is
optimised for code size rather than performance.

It is exposed in the debug print and printf.

This increases code size and so is off by default.  Floats will be
printed as <float> or <double> if this support is not present.

You can enable it with the following two xmake flags:

--debug-print-doubles=y
--debug-print-floats=y

Note that `printf` never prints floats.  The C calling convention
promotes all floats to doubles when passed as variadic arguments.  This
means that debug-print-doubles is the only relevant one for doubles.