sdk/firmware.ldscript: introduce explicit PHDRs
This captures what we were doing already -- the firmware images should
be more or less unaltered except perhaps in choice of padding bytes --
but by forcing things into the same LOAD program header, the section
ALIGN() directives ensure that the things we load are at least 4-byte
aligned and sized.

This matters for OpenOCD talking to the Sonata debug unit, which can
write only aligned multiples of four bytes to memories.