ldscripts: ALIGN does not move .
Prior to this change, it was possible that, for example,
`__compart_cgps` names an address *greater* than `.scheduler_globals`,
because `.` may not be `64`-byte aligned when the former is computed,
and is unchanged while subsequently placing the latter.

That isn't what we meant, but it isn't caught because...

- the only use of `__compart_cgps` is in debug loaders,

- iterating over the `libraries_and_compartments()` of the firmware
  image does not include the privileged compartments and libraries
  (specifically: scheduler, allocator, token library, and optional
  sofwtare revoker), and

- the skipped components cumulatively have more than 64 bytes of state,
  so the erroneous value is guaranteed to be before the non-privileged
  regions that do get iterated over.

Correct this pattern in the few places it shows up.