allocator-tests: improve hazard tests.
Waiting on `heap_quota_remaining(MALLOC_CAPABILITY) < MALLOC_QUOTA` to
check if the lambda was freed is super confusing (especially since the
test does not document why it is doing this). Instead, wait until the
allocator returns `-EINVAL` for the pointers we freed, which indicates
that 1) the ephemeral claims have been released and 2) the lambda was
freed since it called the allocator.

This also has the advantage of documenting the difference between
`heap_free` returning `-EPERM` (the chunk has been freed but is waiting
on the hazard list) and `-EINVAL` (the chunk has actually been freed).

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@ubc.ca>