alloc.h: fix race between claim and ephemeral claim.
There is no check for the number of claims in `mspace_free`, which
causes issues when an object is claimed while in the hazard quarantine.
As soon as it leaves the hazard quarantine, the allocator assumes that
there are no outstanding claims, and proceeds painting and freeing
`ptr`. This may happen if an object is claimed while an ephemeral claim
is holding the allocator from freeing it.

This bug is doubly (triply?) bad: we free an object we said was claimed,
we do not refund the quota associated to this claim, and we leak any
associated claim record object(s), since there would, after free, be no
initial pointer to the list thereof.

This corner case was reported by Reto Achermann (UBC) as part of an
ongoing verification effort of the memory allocator.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@ubc.ca>