allocator-test: remove test for internal implementation detail.
We currently test if the claim size returned by `heap_claim()` is
constrained between the allocation size and allocation size plus the
size of a claim object.

As mentioned by @nwf we do not account for the size of a claim object in
the return value of `heap_claim()`, so `(claimSize <= allocSize +
CHERIOTHeapMinChunkSize)` should always be true.

In addition, CHERIOTHeapMinChunkSize is an implementation detail that
may not be provided by other allocator implementations, such as the
formally-verified Rust allocator we are developing at UBC. This will be
a problem if we want to re-use the test suite.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@ubc.ca>