scheduler: fix overflow in next timer value calculation.
There was a potential overflow of a 32-bit integer in the calculation
of the next timer comparison value if a timeout for large-ish number
of ticks was requested (corresponding to minutes, depending on the
frequency of mtime).  This would have resulted in a shorter timeout
than expected.  Fix this by casting to a 64-bit integer before
multiplying.  Note that the scheduler already assumes that a 64-bit
mtime counter will never overflow.