bench/irq-latency-ext: measure wait and wake times
Report both, in case it's of interest.  Wait time includes all the work
necessary for the scheduler to put the high-priority thread to sleep and
get to the point where IRQs are (checked or) delivered, for the IRQ to
fire and route through the PLIC, and the wake time (i.e., the time
between the core beginning to process the the response time and the
high-priority thread being on core again).

Measuring both lets us fix the Invariant that we're probing at the right
thing, by ensuring that we waited for longer than it took to wake us
(concretely, that the `METRIC()` increases more across the high-priority
thread's `futex_timed_wait()` than it does between the last sample taken
by the low priority thread and the high-priority thread's wake).  If
this isn't true, "wake" times aren't measuring what we mean them to.

While here, fix some clang-tidy complaints.

Thanks to https://github.com/elliotb-lowrisc; see further discussion at
https://github.com/CHERIoT-Platform/cheriot-rtos/issues/586 .