allocator-test: probe multi-issue load barrier
The new CHERIoT-Kudu core has a load barrier per issue pipeline.  Fling
several capability load instructions in a row at the frontend in hopes
of pushing one into each issue pipeline.  All of these should come back
untagged, but in early versions of Kudu there was a typo in the
instantiation of one of the issue pipeline's load barriers, causing it
to miss.