tests: preflight vs clang21 analyser
clang21's clang-tidy correctly spots that our preflight test of the
CHERI::PermissionSet Iterator does not ensure that its access into the
array of expected iterated values is in bounds.  Add an explicit test
that it can use to refine its analysis (and which should make failures
easier for humans to understand).