Sealed objects: make sealed address the payload's
This means that it is now never possible to confuse a sealed handle for
a pointer to the entire allocation, as its address is always past the
sealed object header, and thus interior to the allocation.

While here, shuffle some type definitions out to their own file for
reuse in the loader so we don't need mysterious constants.