Updated `heap_can_free` documentation to reflect return values
The comments now reflect the possible error values it can return.
Specifically, we return `-EINVAL` if `ptr` is not a valid pointer
into a live heap allocation, and `-EPERM` if `heapCapability` is
not valid or it cannot free the provided allocation. This can
happen when a non-owner quota tries to free an allocation for
which it has no claim.