Changed `heap_quota_remaining` return value to align with convention
`heap_quota_remaining` now returns `-EPERM` instead of `-1` when a null
heap cap is provided or unseal returns null, reflecting the behaviour of
other allocator APIs. Also ammended the description to reflect the fact
that `-ENOTENOUGHSTACK` is returned when there isn't enough stack to
execute the function.

Also added a test to ensure a call to `heap_quota_remaining` returns the
correct value.

This does not affect functionality, as in practice this value is
just compared against 0 to verify success.