Updated heap_claim return value to `-ENOMEM` in case of failed call to heap_claim.
`-ENOMEM` better reflects the failure modes of `heap_claim`, which can
happen when a quota is not big enough to accomodate a claim, or the
actual call to Claim::create fails due to a lack of memory, which isn't
described well by `-EPERM`.

Changes applied across stdlib documentation, implementation, and test
suites.