loader: incrementally link into one .o first
This makes it easier to see what the external dependencies of the loader
actually are (use nm on the resulting file) and to ensure that the
outside world doesn't accidentally come to depend on anything within.

Fib a little bit to xmake by saying we're making a "binary"-kinded
target (so we get both compilation and link phases, passing ldflags to
make the latter incremental) and rely on the target file extension to
convey our intent to the linker.