loader: incrementally link into one .o first
This makes it easier to see what the external dependencies of the loader
actually are (use nm on the resulting file) and to ensure that the
outside world doesn't accidentally come to depend on anything within.

Unhook the loader from the "cheriot.component" rule and its custom
linker invocation in favor of xmake's default handling.  Fib a little
bit and say we're making a "binary"-kinded target (so we get both
compilation and link phases, using the dynamic linker) and rely on the
target file extension to convey our intent to the linker.