Temporary fix for failure in `test_claims` after `test_token_hazards`.
See Issue #509. If the teardown code for the closure created as part of
the async invocation in `test_token_hazards` is pre-empted before it
enters `token_obj_destroy()`, `MALLOC_CAPABILITY` is not refunded its
quota, leading to failures in later test cases that check for leaked
objects by testing `heap_quota_remaining`.