heap_free and heap_claim: test, document and fix some edge cases.
1) Use the capability base to look up the chunk to heap_free instead
of the address. This guarantees the address is within the bounds of
the original allocation, unlike the address. Same goes for heap_claim.

2) Test that an attempt to free with a 'one-past-the-end' zero length
capability fails.

3) Test that an attempt to free fails if the bounds are not the full
allocation.

4) Test that it's OK to free a pointer to the interior of an
allocation if it still has the full bounds.

5) Don't allow claim or free of sealed capabilities: there's no reason
to claim a sealed capability and for free the correct API is to use
token_obj_destroy with the correct sealing key. Add the necessary
checks and tests for this. Note that freeing fails for a reason when
using the unsealed handle to a sealed token: the unsealed handle does
not encompass the whole allocation (including header) so the free
fails with a different error code. Note that taking a claim to a
suballocation and subsequently releasing it using heap_free is
allowed, so we can do that with the unsealed capability to a sealed
token.