Stop ephemeral claims from working on sealed capabilities.
We don't allow heap_claim on sealed capabilities so it probably
shouldn't work for ephemeral claims either.  The check has to be done
in the allocator and not heap_claim_ephemeral because as a libcall it
is not a security boundary.  Note that the testsubset instruction does
not check for sealed capabilities.  I'm not sure whether we should
also check for this there and return -EINVAL.

Also added a test.